feat: new-ui look
