+++ lib/components/channel_panel.dart
@@ -2,22 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:rtchat/components/chat_panel.dart';
 import 'package:rtchat/components/emote_picker.dart';
-import 'package:rtchat/components/statistics_bar.dart';
 import 'package:rtchat/models/adapters/actions.dart';
 import 'package:rtchat/models/channels.dart';
 import 'package:rtchat/models/layout.dart';
 import 'package:rtchat/models/tts.dart';
 import 'package:rtchat/models/commands.dart';
 import 'package:flutter_keyboard_visibility/flutter_keyboard_visibility.dart';
-
-import 'channel_search_dialog.dart';
-
-class _ChannelPickerValue {
-  final Channel? channel;
-  final bool isAdd;
-
-  const _ChannelPickerValue({this.channel, this.isAdd = false});
-}
 
 class ChannelPanelWidget extends StatefulWidget {
   final void Function(bool)? onScrollback;
@@ -69,103 +59,27 @@ class _ChannelPanelWidgetState extends State<ChannelPanelWidget> {
           }
           widget.onResize!(details.delta.dy);
         },
-        child: Row(children: [
-          Consumer<LayoutModel>(builder: (context, layoutModel, child) {
-            if (layoutModel.locked) {
-              return Container();
-            }
-            return const Padding(
-                padding: EdgeInsets.only(right: 16),
-                child: Icon(Icons.drag_indicator));
-          }),
-          Consumer<ChannelsModel>(builder: (context, channelsModel, child) {
-            if (channelsModel.subscribedChannels.isEmpty) {
-              return const Spacer();
-            }
-            final first = channelsModel.subscribedChannels.first;
-            return Expanded(
-              child: PopupMenuButton<_ChannelPickerValue>(
-                  child: Row(
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      children: [
-                        Padding(
-                          padding: const EdgeInsets.only(right: 8),
-                          child: Image(
-                              height: 24,
-                              image: AssetImage(
-                                  'assets/providers/${first.provider}.png')),
-                        ),
-                        Expanded(
-                          child: Text("/${first.displayName}",
-                              softWrap: false, overflow: TextOverflow.fade),
-                        ),
-                      ]),
-                  onSelected: (value) async {
-                    if (value.isAdd) {
-                      // show the search dialog.
-                      await showDialog(
-                        context: context,
-                        builder: (context) =>
-                            ChannelSearchDialog(onSelect: (channel) {
-                          channelsModel.subscribedChannels = {channel};
-                          Navigator.pop(context);
-                        }),
-                      );
-                    } else {
-                      channelsModel.subscribedChannels = {value.channel!};
-                    }
-                  },
-                  itemBuilder: (context) {
-                    return [
-                      ...channelsModel.availableChannels.map((channel) {
-                        return PopupMenuItem(
-                            value: _ChannelPickerValue(channel: channel),
-                            child: ListTile(
-                              title: Text(channel.displayName),
-                            ));
-                      }),
-                      const PopupMenuItem(
-                          value: _ChannelPickerValue(isAdd: true),
-                          child: ListTile(
-                              title: Text("Find a channel"),
-                              leading: Icon(Icons.add)))
-                    ];
-                  }),
-            );
-          }),
-          Consumer<ChannelsModel>(builder: (context, channelsModel, child) {
-            if (channelsModel.subscribedChannels.isEmpty) {
-              return Container();
-            }
-            final first = channelsModel.subscribedChannels.first;
-            return StatisticsBarWidget(
-                provider: first.provider, channelId: first.channelId);
-          }),
-          Consumer<TtsModel>(builder: (context, ttsModel, child) {
-            return IconButton(
-                icon: Icon(ttsModel.enabled
-                    ? Icons.record_voice_over
-                    : Icons.voice_over_off),
-                tooltip: "Text to speech",
-                onPressed: () {
-                  ttsModel.enabled = !ttsModel.enabled;
-                });
-          }),
-        ]),
+        child: Consumer<LayoutModel>(builder: (context, layoutModel, child) {
+          if (layoutModel.locked) {
+            return Container();
+          }
+          return Center(
+            child: ClipRRect(
+              borderRadius: BorderRadius.circular(15.0),
+              child: Container(
+                width: 350,
+                height: 16,
+                color: Colors.blueAccent,
+              ),
+            ),
+          );
+        }),
       ),
     );
 
     return Column(children: [
-      // header
-      DefaultTextStyle.merge(
-          style: Theme.of(context).primaryTextTheme.subtitle2,
-          child: IconTheme(
-              data: Theme.of(context).primaryIconTheme,
-              child: Container(
-                  height: 56,
-                  color: Theme.of(context).primaryColor,
-                  child: header))),
+      // drag bar indicator,
+      header,
 
       // body
       Expanded(
@@ -254,6 +168,17 @@ class _ChannelPanelWidgetState extends State<ChannelPanelWidget> {
             //     }).toList();
             //   },
             // ),
+            Consumer<TtsModel>(builder: (context, ttsModel, child) {
+              return IconButton(
+                icon: Icon(ttsModel.enabled
+                    ? Icons.record_voice_over
+                    : Icons.voice_over_off),
+                tooltip: "Text to speech",
+                onPressed: () {
+                  ttsModel.enabled = !ttsModel.enabled;
+                },
+              );
+            }),
             _buildSendButton(),
           ]),
         );
