+++ lib/screens/settings/settings.dart
@@ -1,11 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:provider/provider.dart';
+import 'package:rtchat/components/channel_search_dialog.dart';
+import 'package:rtchat/models/channels.dart';
 import 'package:rtchat/models/layout.dart';
 import 'package:rtchat/models/style.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 const discordUrl = "https://discord.gg/UKHJMQs74u";
+
+class _ChannelPickerValue {
+  final Channel? channel;
+  final bool isAdd;
+
+  const _ChannelPickerValue({this.channel, this.isAdd = false});
+}
 
 Widget _iconWithText(IconData icon, String text) {
   return Column(children: [
@@ -33,6 +42,65 @@ class _SettingsScreenState extends State<SettingsScreen> {
       ),
       body: Consumer<LayoutModel>(builder: (context, layoutModel, child) {
         return ListView(children: [
+          ListTile(
+            title: Consumer<ChannelsModel>(
+                builder: (context, channelsModel, child) {
+              if (channelsModel.subscribedChannels.isEmpty) {
+                return const Spacer();
+              }
+              final first = channelsModel.subscribedChannels.first;
+              return Expanded(
+                child: PopupMenuButton<_ChannelPickerValue>(
+                    child: Row(
+                        mainAxisAlignment: MainAxisAlignment.start,
+                        crossAxisAlignment: CrossAxisAlignment.center,
+                        children: [
+                          Padding(
+                            padding: const EdgeInsets.only(right: 8),
+                            child: Image(
+                                height: 24,
+                                image: AssetImage(
+                                    'assets/providers/${first.provider}.png')),
+                          ),
+                          Expanded(
+                            child: Text("/${first.displayName}",
+                                softWrap: false, overflow: TextOverflow.fade),
+                          ),
+                        ]),
+                    onSelected: (value) async {
+                      if (value.isAdd) {
+                        // show the search dialog.
+                        await showDialog(
+                          context: context,
+                          builder: (context) =>
+                              ChannelSearchDialog(onSelect: (channel) {
+                            channelsModel.subscribedChannels = {channel};
+                            Navigator.pop(context);
+                          }),
+                        );
+                      } else {
+                        channelsModel.subscribedChannels = {value.channel!};
+                      }
+                    },
+                    itemBuilder: (context) {
+                      return [
+                        ...channelsModel.availableChannels.map((channel) {
+                          return PopupMenuItem(
+                              value: _ChannelPickerValue(channel: channel),
+                              child: ListTile(
+                                title: Text(channel.displayName),
+                              ));
+                        }),
+                        const PopupMenuItem(
+                            value: _ChannelPickerValue(isAdd: true),
+                            child: ListTile(
+                                title: Text("Find a channel"),
+                                leading: Icon(Icons.add)))
+                      ];
+                    }),
+              );
+            }),
+          ),
           ListTile(
             title: const Text('Activity feed'),
             subtitle: const Text("Customize your activity feed"),
